// src/types.d.ts
export interface Cart {
  _id: string;
  totalItems: number;
  subtotal: number;
  total: number;
  items: CartItem[];
  discounts: {
    code: string;
    amount: number;
  }[];
  currency: {
    code: string;
    symbol: string;
  };
  createdAt:Date
  updatedAt:Date
}

export interface CartItem {
  _id: string;
  productId: string;
  name: string;
  quantity: number;
  price: number;
  subtotal: number;
  total: number;
  variant?: string;
  image?: string;
}